Create week 4 tasks (#3)
* Create week 4 tasks

Completed tasksCompleted.swift, main.swift and tasks.swift. File "tasks after task0.swift" is for reference and is a compilable version of tasks.swift after the changes in Task0 have been completed. Minor corrections to a few other files from earlier weeks. Did not yet implement change to week 2 task 4, nor did make matching changes in week 4 to index.html and ponder_prove.html. Those three changes are next on the list, but I wanted to get this much checked in.

* Week 2 task 4 and week 4 tasks 1 and 2

Modified week 2 task 4 as suggested.

Made one change to week 4 tasks 1 and 2, moving the definition of "func ==" from task 2 to task 1 where it makes more sense now in context of having developed the rest of the tasks in week 4.

* Week 4 reading and task adjustments

Adjustments and additions to Week 4 reading in index.html. Rework ponder_prove.html to match the new task process (but Layne needs to fix the links to the two files to download). One addition to the Week 2 reading that came up several times as best practice when using optionals in the tasks. Optional Chaining is used more in later weeks, but the best place to explain it is in week 2 with the introduction of Optionals and some built in functions like Filter() and Map.

* Week 5 reading and assignment html

Updated week 5 reading and assignment html.
Also corrected a couple typos in the introductory comments included in all task files. Next step is creation of main.swift and tasksCompleted.swift for week 5, some parts of which are already developed, but that should be a separate branch after this one is closed. Still need to decide if tasks 7-9 from week 3 stay there or move to week 5.

* Correct typos

Correct a couple typos/discrepancies in Week 5 reading heading

* Add draft of Week 5 tasks

Tasks.swift and main.swift implementing tasks 3-9. Skipped tasks 0,1,2 for now. TasksCompleted.swift has completed versions of of those tasks.
Also includes a few corrections to the reading that came up as the tasks were developed.

* Correct print to testPrint

Corrected two typos in task9 sample code. Corrected many cases in sample code where print() was called rather than testPrint(). Did not matter in this case, but need to set the correct example. One small change to the engine part of main.swift to allow testPrint() to accept the terminator: parameter.

* Correct week 3 reading and tasks

Change week 3 reading and tasks to properly use the data model and not create objects from the parent type. Added "tasks after task0.swift" which is a copy of tasks.swift that has had only task0 completed so we have a compilable version of the file without any other tasks done (only for internal validation). Corrected two typos in week 4 files and one in week5. The one in week 5 will be deleted, but was caught and fixed earlier, just not committed.

Co-authored-by: Layne Moseley <laynemoseley@gmail.com>