Insert Task in week 4 (#7)
Inserted new Task 1 in week 4 (revising the function we had them write for the last Task of week 3 to work with the new data model). The function they wrote to work on an array of MedicationContainers now needs to be a method that works on a Dictionary of Sets of MedicationContainers. Renumbered the other tasks. Cleaned and removed up some Task 0 tests. Correct two typos noticed in week 3 files. Next commit is planned to be a final cleanup pass on week 4.