Complete edits to week 4 tasks (#8)
Completed all remaining edits to week 4 tasks and tests. Cleaned up wording, removed some "handholding" and clarified some reasons and results. Task 9 has some handholding to get things set up since this is a new concept, but it still has them code the actual generic function themselves. Adjusted the wording in the reading and assignment html to match the way the tasks actually work and the fact that we are not introducing the PharmaceuticalStockTracker data model for the first time this week. Next steps are cleaning up the parts of week 5 other than the mini-projects (which will be a later step) and adjusting the syllabus to match the changes in assignments and expectations.