Add a draft mini project (#10)
Added a draft mini-project, both incomplete in tasks.swift and a completed implementation in tasksCompleted.swift. Very possibly it is too much work, but we can figure out how to scale it back, for example reducing the number of operators, or removing the binary operators. They might also find checking for circular references to be beyond their skills to do quickly although that was quick for me. Adjusted the enum types to provide nice formatting and comparisons to help testing and debugging. Realized the reading content was missing information about how to pull associated values back out so added a paragraph and sample for that.