Update syllabus.html (#11)
* Update syllabus.html

ON reviewing the Syllabus, I found I could edit parts, but the first part not only talks about teamwork, but other things including the overall approach and discusses how they will see how things they learned in specific prior courses "are used to solve problems in your discipline" which seems copied from another course since that does not really match what we are doing. It felt like someone more familiar with the other courses and overall pedantic strategy and approach needed to edit that section so I labeled it as a section for Layne to edit. Of course he will want to review and edit the other section as well, but at least I felt confident doing some edits there.

* More edits to the syllabus

Co-authored-by: Layne Moseley <laynemoseley@gmail.com>