Correction to Week 5
This contains two corrections to the Week 5 reading that contain leftover references to teaching them about pulling data from a web site. So ideally we would get these posted right away. I took the opportunity to correct a few typos in the file and co clean up the wording in one case.