Urgent fixes to main.swift
The Windows compiler could not properly handle my having the main body of code earlier in the file above where some globals were defined. Moving it down to the bottom of the file solved the problem that was keeping some tasks from compiling. I posted an announcement with the new version for week 4, but it would be good to get the correction for week 5 published. I still owe you to finish editing the typos in the weeks 2 and 3 changes that are outstanding which is why these are a separate commit.