Another fix to week 5
This corrects an error a student found in my code for Task 0. I had missed including the distance when calculating the current position. I corrected it in the completed sample code, but especially in the main.swift where it does the tests. No new version of tasks.swift this time since it was only in the sample code for that one.