Week 4 changes
Correct invalid link in index.html.
Remove tests 2 and 9
Renumber tests 3-8 to be 2-7
Remove use of the Regular Expression based checker that was task 2 in the assignments and sample code for the new tasks 2, 3 and 4.
Revised the explanation in Task 1 which previously referenced week 3 task 10 that is being deleted.
Note that tasks.swift changes are mirrored in tasks after task0 and tasks completed. The small set of changes is replicated in all three files.