Correct Errors
This corrects some errors that caused difficulties for some students. It also provides better wording when students often misinterpreted instructions. For week 5 testPrint() was corrected to generate the correct count of lines if a student used the option to continue lines. It corrects a missing brace that somehow came into the current version of the week 4 task that caused students unnecessary difficulty.