Correct installation and common misunderstandings
This has corrections to the installation to work with new pages from the Swift organization. It also combines the WIndows and Linux parts where they are common. And makes the stretch challenge only for MacOS users since the debugger documentation is not stable for other users. It also addresses some things that caused many users difficulties in past semesters with better explanations or more flexible testing. I also added a few best practice recommendations right in the week 6 assignment in hopes that I will not need to ask as many rewrites and so I can point at them if I need to ask them to revise for best practices.